+++ src/app/compiler/dependencies.ts
@@ -2,15 +2,19 @@ import * as _ from 'lodash';
 import * as path from 'path';
 import * as util from 'util';
 import * as ts from 'typescript';
-import marked from 'marked';
+import { readFileSync } from 'fs-extra';
+
 import { compilerHost, detectIndent } from '../../utilities';
 import { logger } from '../../logger';
 import { RouterParser } from '../../utils/router.parser';
 import { LinkParser } from '../../utils/link-parser';
 import { JSDocTagsParser } from '../../utils/jsdoc.parser';
 import { generate } from './codegen';
+import { stripBom } from '../../utils/utils';
 import { Configuration, IConfiguration } from '../configuration';
 import { $componentsTreeEngine } from '../engines/components-tree.engine';
+
+const marked = require('marked');
 
 interface NodeObject {
     kind: Number;
@@ -126,7 +130,7 @@ export class Dependencies {
                 types: []
             }
         };
-        let sourceFiles = this.program.getSourceFiles() || [];
+        let sourceFiles = this.parseFiles(this.files) || [];
 
         sourceFiles.map((file: ts.SourceFile) => {
 
@@ -170,6 +174,19 @@ export class Dependencies {
         return deps;
     }
 
+    private parseFiles(fileNames: string[]) {
+
+        let _parsedFiles = [];
+
+        fileNames.forEach(fileName => {
+            let sourceCode = stripBom(readFileSync(fileName).toString()),
+                sourceFile = ts.createSourceFile(fileName, sourceCode, ts.ScriptTarget.ES5, true);
+            _parsedFiles.push(sourceFile);
+        });
+
+        return _parsedFiles;
+    }
+
 
     private getSourceFileDecorators(srcFile: ts.SourceFile, outputSymbols: Object): void {
 
@@ -186,7 +203,7 @@ export class Dependencies {
             if (node.decorators) {
                 let visitNode = (visitedNode, index) => {
 
-                    let metadata = node.decorators.pop();
+                    let metadata = node.decorators;
                     let name = this.getSymboleName(node);
                     let props = this.findProps(visitedNode);
                     let IO = this.getComponentIO(file, sourceFile);
@@ -615,24 +632,44 @@ export class Dependencies {
         return result;
     }
 
-    private isComponent(metadata) {
-        return metadata.expression.expression.text === 'Component';
+    private parseDecorators(decorators, type: string): boolean {
+        let result = false;
+        if (decorators.length > 1) {
+            _.forEach(decorators, function(decorator) {
+                if (decorator.expression.expression) {
+                    if (decorator.expression.expression.text === type) {
+                        result = true;
+                    }
+                }
+            });
+        } else {
+            if (decorators[0].expression.expression) {
+                if (decorators[0].expression.expression.text === type) {
+                    result = true;
+                }
+            }
+        }
+        return result;
     }
 
-    private isPipe(metadata) {
-        return metadata.expression.expression.text === 'Pipe';
+    private isComponent(metadatas) {
+        return this.parseDecorators(metadatas, 'Component');
     }
 
-    private isDirective(metadata) {
-        return metadata.expression.expression.text === 'Directive';
+    private isPipe(metadatas) {
+        return this.parseDecorators(metadatas, 'Pipe');
     }
 
-    private isInjectable(metadata) {
-        return metadata.expression.expression.text === 'Injectable';
+    private isDirective(metadatas) {
+        return this.parseDecorators(metadatas, 'Directive');
     }
 
-    private isModule(metadata) {
-        return metadata.expression.expression.text === 'NgModule';
+    private isInjectable(metadatas) {
+        return this.parseDecorators(metadatas, 'Injectable');
+    }
+
+    private isModule(metadatas) {
+        return this.parseDecorators(metadatas, 'NgModule');
     }
 
     private getType(name) {
@@ -964,6 +1001,10 @@ export class Dependencies {
             result.description = marked(LinkParser.resolveLinks(ts.displayPartsToString(method.symbol.getDocumentationComment())));
         }
 
+        if (method.decorators) {
+            result.decorators = this.formatDecorators(method.decorators);
+        }
+
         if (method.modifiers) {
             if (method.modifiers.length > 0) {
                 result.modifierKind = method.modifiers[0].kind;
@@ -1015,6 +1056,33 @@ export class Dependencies {
         }
     }
 
+    private formatDecorators(decorators) {
+        let _decorators = [];
+
+        _.forEach(decorators, (decorator) => {
+            if (decorator.expression) {
+                if (decorator.expression.text) {
+                    _decorators.push({
+                        name: decorator.expression.text
+                    });
+                }
+                if (decorator.expression.expression) {
+                    var info = {
+                        name: decorator.expression.expression.text
+                    }
+                    if (decorator.expression.expression.arguments) {
+                        if (decorator.expression.expression.arguments.length > 0) {
+                            info.args = decorator.expression.expression.arguments;
+                        }
+                    }
+                    _decorators.push(info);
+                }
+            }
+        });
+
+        return _decorators;
+    }
+
     private visitProperty(property, sourceFile) {
         /**
          * Copyright https://github.com/ng-bootstrap/ng-bootstrap
@@ -1027,12 +1095,16 @@ export class Dependencies {
              line: this.getPosition(property, sourceFile).line + 1
          }
 
+         if (property.decorators) {
+             result.decorators = this.formatDecorators(property.decorators);
+         }
+
          if (property.modifiers) {
              if (property.modifiers.length > 0) {
                  result.modifierKind = property.modifiers[0].kind;
              }
          }
-        return result;
+         return result;
     }
 
     private visitMembers(members, sourceFile) {
@@ -1129,32 +1201,24 @@ export class Dependencies {
     }
 
     private isPipeDecorator(decorator) {
-        /**
-         * Copyright https://github.com/ng-bootstrap/ng-bootstrap
-         */
-         return decorator.expression.expression.text === 'Pipe';
+        return (decorator.expression.expression) ? decorator.expression.expression.text === 'Pipe' : false;
     }
 
     private isModuleDecorator(decorator) {
-        /**
-         * Copyright https://github.com/ng-bootstrap/ng-bootstrap
-         */
-         return decorator.expression.expression.text === 'NgModule';
+        return (decorator.expression.expression) ? decorator.expression.expression.text === 'NgModule' : false;
     }
 
     private isDirectiveDecorator(decorator) {
-        /**
-         * Copyright https://github.com/ng-bootstrap/ng-bootstrap
-         */
-        var decoratorIdentifierText = decorator.expression.expression.text;
-        return decoratorIdentifierText === 'Directive' || decoratorIdentifierText === 'Component';
+        if (decorator.expression.expression) {
+            var decoratorIdentifierText = decorator.expression.expression.text;
+            return decoratorIdentifierText === 'Directive' || decoratorIdentifierText === 'Component';
+        } else {
+            return false;
+        }
     }
 
     private isServiceDecorator(decorator) {
-        /**
-         * Copyright https://github.com/ng-bootstrap/ng-bootstrap
-         */
-         return decorator.expression.expression.text === 'Injectable';
+        return (decorator.expression.expression) ? decorator.expression.expression.text === 'Injectable' : false;
     }
 
     private visitClassDeclaration(fileName, classDeclaration, sourceFile?) {
@@ -1215,26 +1279,28 @@ export class Dependencies {
                         implements: implementsElements
                     };
                 } else if (this.isServiceDecorator(classDeclaration.decorators[i])) {
-                  members = this.visitMembers(classDeclaration.members, sourceFile);
-                  return [{
-                    fileName,
-                    className,
-                    description,
-                    methods: members.methods,
-                    indexSignatures: members.indexSignatures,
-                    properties: members.properties,
-                    kind: members.kind,
-                    constructor: members.constructor,
-                    extends: extendsElement,
-                    implements: implementsElements
-                  }];
-              } else if (this.isPipeDecorator(classDeclaration.decorators[i]) || this.isModuleDecorator(classDeclaration.decorators[i])) {
-                  return [{
-                    fileName,
-                    className,
-                    description,
-                    jsdoctags: jsdoctags
-                  }];
+                    members = this.visitMembers(classDeclaration.members, sourceFile);
+                    return [{
+                        fileName,
+                        className,
+                        description,
+                        methods: members.methods,
+                        indexSignatures: members.indexSignatures,
+                        properties: members.properties,
+                        kind: members.kind,
+                        constructor: members.constructor,
+                        extends: extendsElement,
+                        implements: implementsElements
+                    }];
+                } else if (this.isPipeDecorator(classDeclaration.decorators[i]) || this.isModuleDecorator(classDeclaration.decorators[i])) {
+                    return [{
+                        fileName,
+                        className,
+                        description,
+                        jsdoctags: jsdoctags
+                    }];
+                } else {
+                    //console.log('custom decorator');
                 }
             }
         } else if (description) {