+++ src/app/engines/search.engine.ts
@@ -56,12 +56,14 @@ export class SearchEngine {
                            index: JSON.stringify(this.getSearchIndex()),
                            store: JSON.stringify(this.documentsStore)
                        });
+                   outputFolder = outputFolder.replace(process.cwd(), '');
                    fs.outputFile(path.resolve(process.cwd() + path.sep + outputFolder + path.sep + '/js/search/search_index.js'), result, function (err) {
                        if(err) {
                            logger.error('Error during search index file generation ', err);
                            reject(err);
+                       } else {
+                           resolve();
                        }
-                       resolve();
                    });
                }
            });