+++ src/app/interfaces/parsed-data.interface.ts
@@ -1,0 +1,11 @@
+export interface ParsedData {
+    modules;
+    components;
+    directives;
+    injectables;
+    interfaces;
+    pipes;
+    classes;
+    miscellaneous;
+    routesTree;
+}