+++ test/src/cli/cli-components.spec.ts
@@ -9,30 +9,31 @@ const env = Object.freeze({TS_NODE_PROJECT: tsconfigPath, MODE:'TESTING'});
 
 describe('CLI components', () => {
 
-    describe('README file', () => {
-
-        let stdoutString = null, todoComponentFile, listComponentFile;
-        before(function (done) {
-            tmp.create();
-            exec(tsNodePath + ' ./bin/index-cli.js -p ./test/src/todomvc-ng2/src/tsconfig.json -d ' + tmp.name + '/', {env}, (error, stdout, stderr) => {
-                if (error) {
-                    console.error(`exec error: ${error}`);
-                    done('error');
-                    return;
-                }
-                stdoutString = stdout;
-                todoComponentFile = read(`${tmp.name}/components/TodoComponent.html`);
-                listComponentFile = read(`${tmp.name}/components/ListComponent.html`);
-                done();
-            });
+    let stdoutString = null, todoComponentFile, listComponentFile, footerComponentFile;
+    before(function (done) {
+        tmp.create();
+        exec(tsNodePath + ' ./bin/index-cli.js -p ./test/src/todomvc-ng2/src/tsconfig.json -d ' + tmp.name + '/', {env}, (error, stdout, stderr) => {
+            if (error) {
+                console.error(`exec error: ${error}`);
+                done('error');
+                return;
+            }
+            stdoutString = stdout;
+            todoComponentFile = read(`${tmp.name}/components/TodoComponent.html`);
+            footerComponentFile = read(`${tmp.name}/components/FooterComponent.html`);
+            listComponentFile = read(`${tmp.name}/components/ListComponent.html`);
+            done();
         });
-        after(() => tmp.clean());
+    });
+    after(() => tmp.clean());
 
-        it('it should have a readme tab', () => {
-            expect(todoComponentFile).to.contain('readme-tab');
-            expect(listComponentFile).to.contain('readme-tab');
-        });
+    it('it should have a readme tab', () => {
+        expect(todoComponentFile).to.contain('readme-tab');
+        expect(listComponentFile).to.contain('readme-tab');
+    });
 
+    it('it should have a decorator listed', () => {
+        expect(footerComponentFile).to.contain('<i>Decorators : </i><code>LogProperty</code>');
     });
 
 });