+++ test/src/cli/cli.spec.ts
@@ -38,7 +38,7 @@ describe('CLI simple flags', () => {
         afterEach(() => tmp.clean());
 
         it('should display error message', () => {
-            expect(command.stdout.toString()).to.contain('"tsconfig.json" file was not found in the current directory');
+            expect(command.stdout.toString()).to.contain('file was not found in the current directory');
         });
 
         it(`should not create a "documentation" directory`, () => {
@@ -198,28 +198,6 @@ describe('CLI simple flags', () => {
             expect(moduleFile).not.to.contain('components/BarComponent.html');
             expect(moduleFile).not.to.contain('injectables/FooService.html');
             expect(moduleFile).not.to.contain('modules/BarModule.html');
-        });
-    });
-
-    describe('when specific files are excluded in tsconfig', () => {
-        let stdoutString = null;
-        before(function (done) {
-            tmp.create();
-            exec(tsNodePath + ' ./bin/index-cli.js -p ./test/src/sample-files/tsconfig.exclude.json -d ' + tmp.name + '/', {env}, (error, stdout, stderr) => {
-                if (error) {
-                    console.error(`exec error: ${error}`);
-                    done('error');
-                    return;
-                }
-                stdoutString = stdout;
-                done();
-            });
-        });
-        after(() => tmp.clean(tmp.name));
-
-        it('should not create files excluded', () => {
-            const isFileExists = exists(`${tmp.name}/components/BarComponent.js`);
-            expect(isFileExists).to.be.false;
         });
     });