+++ test/src/todomvc-ng2/src/app/footer/footer.component.ts
@@ -4,6 +4,8 @@ import { TodoStore } from '../shared/services/todo.store';
 
 import { EmitterService } from '../shared/services/emitter.service';
 
+import { LogMethod, LogProperty, LogPropertyWithArgs, LogClass } from '../shared/decorators/log.decorator';
+
 /**
  * The footer component
  */
@@ -20,10 +22,13 @@ export class FooterComponent {
     /**
      * Local id for EmitterService
      */
+    @LogProperty
     id: string = 'FooterComponent';
+
     /**
      * Starting filter param
      */
+    @LogPropertyWithArgs('theCurrentFilter')
     currentFilter: string = 'all';
 
     /**
@@ -38,6 +43,7 @@ export class FooterComponent {
     /**
      * Removes all the completed todos
      */
+    @LogMethod
     removeCompleted() {
 		this.todoStore.removeCompleted();
 	}