+++ test/src/todomvc-ng2/src/app/shared/decorators/log.decorator.ts
@@ -1,0 +1,15 @@
+export function LogMethod(target: any, key: string) {
+    console.log('LogMethod: ' + key);
+};
+
+export function LogProperty(target: any, key: string) {
+    console.log('LogProperty: ', key);
+}
+
+export function LogPropertyWithArgs(alias: string): any {
+    console.log('LogPropertyWithArgs: ', alias);
+}
+
+export function LogClass(target: any) {
+    console.log('LogClass: ', target);
+}