+++ test/src/todomvc-ng2/src/app/shared/models/todo.model.ts
@@ -1,4 +1,4 @@
-export class Tada {
+﻿export class Tada {
 
 }